+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ListPage from './pages/MyOpenFunding/ListPage';
+import ListPage from './pages/MyJoinFunding/ListPage';
 const App = () => {
   return (
     <div className='App'>
